fase-3: http4k gets a data layer, and four JVM projects that never compiled
kotlin/http4k answered /db/simple, /db/complex and /cache with hardcoded
literals while declaring no PostgreSQL or Redis dependency; the previous pass
made them return 501 so the runner would record "not implemented" instead of a
fabricated win. They now run the same queries as everyone else, through HikariCP
sized by DB_POOL_MAX and Lettuce, at the versions src/kotlin/ktor uses, with the
normative SQL from the contract.

Its /health also returned {status, version}. The gate checks the key set for
/health, which is {status, version, timestamp, database, cache}, so this
implementation would have failed it.

The rest of this commit exists because there was finally a way to check. There
was no Maven, Gradle or kotlinc on this machine, so the JVM evidence had been
`javac` over a single extracted file, and the plan said so. Gradle 8.5, Maven
3.9.9 and JDKs 21 and 17 were installed into the session's temp directory and
the projects were built. Four did not compile:

  kotlin/http4k       `val JSON = CONTENT_TYPE of APPLICATION_JSON` with no type
                      annotation -- `of` returns a modifier generic in the
                      message type and nothing could infer it
  kotlin/ktor         JsonRoutes.kt never imported io.ktor.server.application,
                      so `call` did not resolve; and the /cache TTL was an Int
                      where getOrSet takes a Long
  java/spring         JsonController.java had doubled braces, {{ and }}, left by
  graalvm/spring      a str.format template that was never unescaped. Plain
  graalvm/gspring     syntax errors, in all three copies
  kotlin/grpc/armeria an allOpen { } block with no kotlin("plugin.allopen")
                      applied: the build script itself did not compile

graalvm/vertx is the interesting one. Its POM declared no dependencies at all --
no Vert.x, no PostgreSQL, no Redis -- and its Dockerfile ran javac over a single
file with no classpath. That is why the old Main was a com.sun.net.httpserver
server answering with invented data: it was the only thing that could compile
there. Dependencies and a Maven build are in; it now fails on named Vert.x API
errors instead, which belongs to Fase 6.

The measurement also corrects Fase 6 downward: of the six Kotlin gRPC/GraphQL
projects listed as broken, kotlin/graphql/{dgs,graphql-kotlin,spring-graphql}
build clean. kotlin/grpc/spring-grpc asks for io.grpc:grpc-spring-boot-starter,
a coordinate that does not exist -- the published one is net.devh -- so it could
never have resolved.

Also: .gitignore covered Maven's target/ but not Gradle's build/ and .gradle/.

Verification: gradle build clean on kotlin/{http4k,ktor,spring} and
kotlin/graphql/{dgs,graphql-kotlin,spring-graphql}; mvn package clean on
java/{spring,micronaut}; mvn compile clean on graalvm/{spring,gspring,micronaut,
helidon}. java/quarkus is inconclusive -- its uber-jar step throws
ReadOnlyFileSystemException, which looks like friction with the network share
the repository is mounted on rather than a code defect.

+++ src/kotlin/ktor/src/main/kotlin/com/benchmark/routes/CacheRoutes.kt
@@ -26,7 +26,7 @@ fun Route.cacheRoutes(cacheService: CacheService) {
     get("/cache") {
         val key = call.request.queryParameters["key"] ?: "test"
         val newValue = "cached-value-${UUID.randomUUID()}"
-        val value = cacheService.getOrSet(key, newValue, CACHE_TTL_SECONDS)
+        val value = cacheService.getOrSet(key, newValue, CACHE_TTL_SECONDS.toLong())
         // The contract carries a boolean plus the TTL, not a free-form
         // "source" string.
         call.respond(
